==> w9_ML_project/notebooks/notebook_01_cleaning.ipynb <==
# table of contents
1. [intro](#intro)
2. [dataset](#dataset)
3. [dictionary](#dictionary)
4. [exploration](#exploration)
5. [cleaning](#cleaning)
6. [statistical summary](#statistical-summary)

# intro
A chain of hotels has just hired me as a data analyst. They have noticed that the cancellation rate has risen in the past few years. This often leads to rooms that are left unrented for multiple days at a time.

Hotel management is interested in developing a model to predict the likelihood that a customer will cancel their reservation. If successful, this could be used to optimize their booking service and anticipate when cancellations will occur.

I'm going to prepare a report that is accessible to a broad audience. It should outline my motivation, steps, findings, and conclusions.

# dataset
This dataset consists of booking data from a city hotel and a resort hotel.\
It includes many details about the bookings, including room specifications, the length of stay, the time between the booking and the stay, whether the booking was canceled, and how the booking was made.

# dictionary
_Note: For binary variables: `1` = true and `0` = false._

| Column                                                                                                                                                                                                          | Explanation                                                                                                                            |   |   |   |
|-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|----------------------------------------------------------------------------------------------------------------------------------------|---|---|---|
| is_canceled                                                                                                                                                                                                     | Binary variable indicating whether a booking was canceled                                                                              |   |   |   |
| lead_time                                                                                                                                                                                                       | Number of days between booking date and arrival date                                                                                   |   |   |   |
| arrival_date_week_number, arrival_date_day_of_month, arrival_date_month                                                                                                                                         | Week number, day date, and month number of arrival date                                                                                |   |   |   |
| stays_in_weekend_nights, stays_in_week_nights                                                                                                                                                                   | Number of weekend nights (Saturday and Sunday) and weeknights (Monday to Friday) the customer booked                                   |   |   |   |
| adults, children, babies                                                                                                                                                                                        | Number of adults, children, babies booked for the stay                                                                                 |   |   |   |
| is_repeated_guest                                                                                                                                                                                               | Binary variable indicating whether the customer was a repeat guest                                                                     |   |   |   |
| previous_cancellations                                                                                                                                                                                          | Number of prior bookings that were canceled by the customer                                                                            |   |   |   |
| previous_bookings_not_canceled                                                                                                                                                                                  | Number of prior bookings that were not canceled by the customer                                                                        |   |   |   |
| required_car_parking_spaces                                                                                                                                                                                     | Number of parking spaces requested by the customer                                                                                     |   |   |   |
| total_of_special_requests                                                                                                                                                                                       | Number of special requests made by the customer                                                                                        |   |   |   |
| avg_daily_rate                                                                                                                                                                                                  | Average daily rate, as defined by dividing the sum of all lodging transactions by the total number of staying nights                   |   |   |   |
| booked_by_company                                                                                                                                                                                               | Binary variable indicating whether a company booked the booking                                                                        |   |   |   |
| booked_by_agent                                                                                                                                                                                                 | Binary variable indicating whether an agent booked the booking                                                                         |   |   |   |
| hotel_City                                                                                                                                                                                                      | Binary variable indicating whether the booked hotel is a "City Hotel"                                                                  |   |   |   |
| hotel_Resort                                                                                                                                                                                                    | Binary variable indicating whether the booked hotel is a "Resort Hotel"                                                                |   |   |   |
| meal_BB                                                                                                                                                                                                         | Binary variable indicating whether a bed & breakfast meal was booked                                                                   |   |   |   |
| meal_HB                                                                                                                                                                                                         | Binary variable indicating whether a half board meal was booked                                                                        |   |   |   |
| meal_FB                                                                                                                                                                                                         | Binary variable indicating whether a full board meal was booked                                                                        |   |   |   |
| meal_No_meal                                                                                                                                                                                                    | Binary variable indicating whether there was no meal package booked                                                                    |   |   |   |
| market_segment_Aviation, market_segment_Complementary, market_segment_Corporate, market_segment_Direct, market_segment_Groups, market_segment_Offline_TA_TO, market_segment_Online_TA, market_segment_Undefined | Indicates market segment designation with a value of 1. "TA"= travel agent, "TO"= tour operators                                       |   |   |   |
| distribution_channel_Corporate, distribution_channel_Direct, distribution_channel_GDS, distribution_channel_TA_TO, distribution_channel_Undefined                                                               | Indicates booking distribution channel with a value of 1. "TA"= travel agent, "TO"= tour operators, "GDS" = Global Distribution System |   |   |   |
| reserved_room_type_A, reserved_room_type_B, reserved_room_type_C, reserved_room_type_D, reserved_room_type_E, reserved_room_type_F, reserved_room_type_G, reserved_room_type_H, reserved_room_type_L            | Indicates code of room type reserved with a value of 1. Code is presented instead of designation for anonymity reasons                 |   |   |   |
| deposit_type_No_Deposit                                                                                                                                                                                         | Binary variable indicating whether a deposit was made                                                                                  |   |   |   |
| deposit_type_Non_Refund                                                                                                                                                                                         | Binary variable indicating whether a deposit was made in the value of the total stay cost                                              |   |   |   |
| deposit_type_Refundable                                                                                                                                                                                         | Binary variable indicating whether a deposit was made with a value under the total stay cost                                           |   |   |   |
| customer_type_Contract                                                                                                                                                                                          | Binary variable indicating whether the booking has an allotment or other type of contract associated to it                             |   |   |   |
| customer_type_Group                                                                                                                                                                                             | Binary variable indicating whether the booking is associated to a group                                                                |   |   |   |
| customer_type_Transient                                                                                                                                                                                         | Binary variable indicating whether the booking is not part of a group or contract, and is not associated to other transient booking    |   |   |   |
| customer_type_Transient-Party                                                                                                                                                                                   | Binary variable indicating whether the booking is transient, but is associated to at least another transient booking                   |   |   |   |

[Source](https://www.kaggle.com/jessemostipak/hotel-booking-demand/) and [license](https://creativecommons.org/licenses/by/4.0/) of data. 

`Citation`: The data is originally from an article called [Hotel booking demand datasets](https://www.sciencedirect.com/science/article/pii/S2352340918315191) by Nuno Antonio, Ana de Almeida, and Luis Nunes. It was cleaned by Thomas Mock and Antoine Bichat for [#TidyTuesday during the week of February 11th, 2020](https://github.com/rfordatascience/tidytuesday/blob/master/data/2020/2020-02-11/readme.md).

# exploration

import pandas as pd
df = pd.read_csv("../data/01_raw_data.csv")
print(df.shape)
display(df.head(61))
display(df.tail(61))

"""
By default, panda methods head and tail give preview of maximum 20 columns, 10 from each side of a dataframe.
The current df has 53 columns, so the default preview doesn't suffice.
The following setting should adjust this and return the preview of all the 53 columns.
"""
pd.set_option('display.max_columns', None)
display(df.head(61))
display(df.tail(61))

# "children" values look like floats. is dtype float? why?
df.info()

# cleaning

#df["children"] = df["children"].astype(int)
#df
# Attempt to immediately convert "children" column dtype to integer (before I forget) fails, due to NaN values.
# So, let's deal with the NaNs.

df.isna().sum()

# "children" is the only column with NaN values, totally 4.

# 4 NaN values / 119210 entries is a negligent proportion, so whatever we do with NaN is irrelevant. Let's drop them.

df = df.dropna()
df

for column in df.columns:
    print(df[column].value_counts())
    print()

# Drop the columns which give no information.
df = df.drop(["market_segment_Undefined", "distribution_channel_Undefined"], axis=1)
df

df["children"] = df["children"].astype(int)
df.info()

# Now all the data is in an appropriate type. Let's check for duplicates.

df.duplicated().any()
#There are duplicates. How many?

df.duplicated().sum()
# There are 36024 duplicates. Where are they?

df[df.duplicated(keep=False)]

1/3 of rows (45039/119210) in the df are duplicated.\
Is this a problem, though? Maybe, but it's hard to know.\
This df is expected to have duplicated rows.\
More than half of the columns are categorical variables with binary values (1,0).\
Out of 119210 entries/guests, it's normal that many guests had identical preferences and booking choices.\
So, what to do with these duplicates? Nothing.

# statistical summary

df.describe()

Of `119206 bookings`:
- `37% were canceled`
- rooms were `booked 2 (median) - 3,5 (mean) months ahead`
  - median < mean = more people tend to book closer to the stay-date
  - but some far ahead bookings (2 years) drag the distribution tail to the right
- `peak arrivals` are `mid-june to mid-july`, showing by the:
  - arrival_date_week_number: 27 (mean) - 28 (median),
  - arrival_date_day_of_month: 15 
  - arrival_date_month: 6.6 (mean) - 7 (median)
- bookings include at least `1 weekend day and/or 2 weekdays`
- bookings were made for `2 adults`, no children or babies
- `3% repeated guests`
- `no history of cancelations`, which aligns with 97% being first-timers
- `car parking space not required`
- `no special requests`
- `daily rate: 95 (median) - 102 (mean)`
- `86% agent` bookings, 6% company bookings
- `66% city hotel` bookings, 34% resort bookings
- meals: `77% bed&breakfast`, 12% half-breakfast, 10% no meal, 1% full breakfast
- market segment: `47% online travel agent (TA)`, 20% offline TA, 17% groups, 11% direct, 4% corporate, 1% other segments
- distribution channel: `82% TA`, 12% direct, 5% corporate
- `72% reserved rooms were type A`, 16% D, 5% E, 7% other 6 types
- `88% booked without deposit`, 12% with a deposit
- `75% individual (transient) bookings`, 21% transient and associated to another transient booking, 4% other arrangements.

`Outliers:`
- 2 years lead time
- 19 weekends
- 50 days
- 55 adults
- 10 children
- 10 babies
- 26 cancelations
- 8 car parking spaces
- 5 special requests
-daily rate: -6.38
- daily rate: 5400

df.to_csv("../data/02_cleaned_data.csv", index=False)

Next: notebook_02_univariate_analysis
